+++ src/game/cRandomGirl.cpp
@@ -251,9 +251,17 @@ std::shared_ptr<sGirl> cRandomGirls::spawn(SpawnReason reason, int age) {
     if(!attempt) attempt = get_spec(reason);
 
     auto new_girl = std::make_shared<sGirl>(false);
+
+    // last call -- just accept any girl
+    if(!attempt && !m_RandomGirls.empty()) {
+        attempt = &m_RandomGirls[g_Dice.random(m_RandomGirls.size())];
+        g_LogFile.warning("girl", "Could not find a fitting random girl template! Spawn reason: ", get_spawn_name(reason));
+    }
+
     if(!attempt) {
         new_girl->m_Desc = "Hard Coded Random Girl\n(The game did not find a valid .rgirlsx file)";
-        new_girl->m_Name = "Default";
+        new_girl->SetName("Default", "", "Girl");
+        g_Game->error("Could not find a valid random girl template!");
         g_LogFile.error("girl", "Could not find a valid random girl template!");
         return new_girl;
     }